Properly handle delete_message() after COW.
When freeing a message payload, do not always assume that the payload pages
are still shared.  The recipient may have modified the payload, in which case
the COW handler will replace the (shared) page frame.  When freeing the
payload buffer then, the new frame will not appear in the shared pool, but
still must be freed to allow subsequent messages to be mapped.