Reworked PAGE_COUNT() to account for misaligned buffers.
The logic assumed that the underlying buffer was always page-aligned; and the
caller was responsible for handling any additional/fractional page.  Just use
PAGE_ALIGN() and PAGE_BASE() instead to compute the exact number of pages
involved, even if the buffer is misaligned or not an integral number of pages.

Fixes an unexpected page fault in the debug build