The 'shell' now is essentially the lua interpreter.
All of the actual 'shell' logic is contained within shell.lua; the executable
now is (mostly) just a standalone lua interpreter.  So start moving towards
a separate interpreter/script model.  This also means that the old 'shell' is
no longer a boot server, so it must be launched explicitly at startup.